wishlist controller logic added

+++ packages/server/src/controllers/event/remove.ts
@@ -1,9 +1,10 @@
 import { eventSchema, type EventForMember } from '@server/entities/event'
 import provideRepos from '@server/trpc/provideRepos'
 import { eventRepository } from '@server/repositories/eventRepository'
-import { publicProcedure } from '@server/trpc'
+import { authenticatedProcedure } from '@server/trpc/authenticatedProcedure'
+import { TRPCError } from '@trpc/server'
 
-export default publicProcedure
+export default authenticatedProcedure
   .use(
     provideRepos({
       eventRepository,
@@ -14,12 +15,26 @@ export default publicProcedure
       id: true,
     })
   )
-  .query(
+  .mutation(
     async ({
       input: { id },
-      ctx: { repos },
+      ctx: { repos, authUser },
     }): Promise<EventForMember> => {
-      const event = await repos.eventRepository.remove(id)
-      return event
+      const event = await repos.eventRepository.find(id)
+      if (!event) {
+        throw new TRPCError({
+          code: 'NOT_FOUND',
+          message: 'Event not found',
+        })
+      }
+      // add for admin too
+      if (event.createdBy !== authUser.id) { 
+        throw new TRPCError({
+          code: 'FORBIDDEN',
+          message: 'Not authorized to delete this event',
+        })
+      }
+
+      return repos.eventRepository.remove(id)
     }
   )
